move gpt utils, add topic_selection to Session
